+++ app/src/main/java/com/example/kotlinapp/ui/pokemoninfo/AboutCard.kt
@@ -1,0 +1,29 @@
+package com.example.kotlinapp.ui.pokemoninfo
+
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.unit.dp
+import com.example.kotlinapp.data.Pokemon
+
+@Composable
+fun AboutCard(modifier: Modifier, pokemon: Pokemon) {
+    Card(
+        modifier = modifier.then(
+            Modifier
+                .fillMaxWidth()
+                .padding(
+                    start = 8.dp,
+                    end = 8.dp,
+                    top = 230.dp,
+                    bottom = 8.dp
+                )
+        ),
+        elevation = CardDefaults.cardElevation(2.dp)
+    ) {
+        AboutCardContent(pokemon)
+    }
+}
