+++ app/src/main/java/com/example/kotlinapp/ui/pokemoninfo/PokemonInfoViewModelFactory.kt
@@ -1,13 +1,12 @@
-//package com.example.kotlinapp.ui.pokemoninfo
-//
-//import androidx.lifecycle.ViewModel
-//import androidx.lifecycle.ViewModelProvider
-//import androidx.lifecycle.createSavedStateHandle
-//import androidx.lifecycle.viewmodel.CreationExtras
-//
-//class PokemonInfoViewModelFactory : ViewModelProvider.Factory {
-//    override fun <T : ViewModel> create(modelClass: Class<T>, extras: CreationExtras): T {
-//        val handle = extras.createSavedStateHandle()
-//        return PokemonInfoViewModel(handle) as T
-//    }
-//}
+package com.example.kotlinapp.ui.pokemoninfo
+
+import androidx.lifecycle.ViewModel
+import androidx.lifecycle.ViewModelProvider
+import androidx.lifecycle.createSavedStateHandle
+import androidx.lifecycle.viewmodel.CreationExtras
+
+class PokemonInfoViewModelFactory(val name: String) : ViewModelProvider.Factory {
+    override fun <T : ViewModel> create(modelClass: Class<T>, extras: CreationExtras): T {
+        return PokemonInfoViewModel(name) as T
+    }
+}
